% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRC_abd}
\alias{CRC_abd}
\title{Metagenomics studies of colorectal cancer (CRC)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 267 rows and 572 columns.
}
\source{
<https://pubmed.ncbi.nlm.nih.gov/30936547/>
}
\usage{
data(CRC_abd, meta)
}
\description{
The "CRC_abd" is a feature-by-sample matrix of species-level relative abundance counts from the five studies.
For the easy of demonstration, we include 267 species under order `Clostridiales`.
The "meta" is a data frame including the sample-level variables from the five studies.
}
\references{
Wirbel, Jakob et al. Nat Med. 2019 Apr;25(4):679-689.
}
\keyword{datasets}
