% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRC_data.R
\docType{data}
\name{CRC_data}
\alias{CRC_data}
\title{Datasets from two metagenomics studies of colorectal cancer (CRC)}
\format{
An object of class \code{list} of length 2.
}
\source{
<https://github.com/zellerlab/crc_meta>
}
\usage{
data(CRC_data)
}
\description{
The "CRC_data" includes "CRC_abd" and "CRC_meta". "CRC_abd" is a sample-by-feature matrix of
relative abundance counts from the two studies including 267 species under order "Clostridiales".
The "CRC_meta" is a data frame including the sample-level variables from the two studies.
}
\examples{
\donttest{
library("miMeta")
data("CRC_data")
CRC_abd <- CRC_data$CRC_abd
CRC_meta <- CRC_data$CRC_meta
}

}
\references{
Wirbel, Jakob et al. Nat Med. 2019 Apr;25(4):679-689.
}
\keyword{datasets}
