% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melody.get.summary.R
\name{melody.get.summary}
\alias{melody.get.summary}
\title{Generate Melody summary statistics for individual studies.}
\usage{
melody.get.summary(
  null.obj,
  covariate.interest,
  cluster = NULL,
  parallel.core = NULL,
  verbose = FALSE
)
}
\arguments{
\item{null.obj}{The output of function "melody.null.model".}

\item{...}{See function melody. In covariate.interest and cluster,
the order of samples should be matched with the order in rel.abd used in the "melody.null.model" function.}
}
\value{
Output a list with each component for a study. The component includes the following elements.
\item{ref}{Reference feature ID for the study.}
\item{est}{A matrix contains the relative-abundance association coefficient estimates for the study. The row names are microbial feature IDs
and the column names are the covariates of interest IDs.}
\item{var}{A matrix contains the variances of the relative-abundance association coefficient estimates for the study.
The row names are microbial feature IDs
and the column names are the covariates of interest IDs.}
\item{n}{Sample size for the study.}
}
\description{
This function directly takes the output object from the "melody.null.model" function and
constructs summary statistics for covariates of interest in each study.
The output can be directly used by the function "melody.meta.summary" to perform meta-analysis.
}
\examples{
\donttest{
library("miMeta")
data("CRC_data")
CRC_abd <- CRC_data$CRC_abd
CRC_meta <- CRC_data$CRC_meta

########## Generate summary statistics ##########
rel.abd <- list()
covariate.interest <- list()
for(d in unique(CRC_meta$Study)){
  rel.abd[[d]] <- CRC_abd[CRC_meta$Sample_ID[CRC_meta$Study == d],]
  disease <- as.numeric(CRC_meta$Group[CRC_meta$Study == d] == "CRC")
  names(disease) <- CRC_meta$Sample_ID[CRC_meta$Study == d]
  covariate.interest[[d]] <- matrix(disease, ncol = 1, dimnames = list(names(disease), "disease"))
}

 null.obj <- melody.null.model(rel.abd = rel.abd, ref = "Coprococcus catus [ref_mOTU_v2_4874]")

 summary.stats <- melody.get.summary(null.obj = null.obj, covariate.interest = covariate.interest)
}

}
\seealso{
\code{\link{melody.null.model}},
\code{\link{melody.meta.summary}},
\code{\link{melody}}
}
