% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melody.merge.summary.R
\name{melody.merge.summary}
\alias{melody.merge.summary}
\title{Merge summary statistics across studies.}
\usage{
melody.merge.summary(melody.obj.lst, verbose = FALSE)
}
\arguments{
\item{melody.obj.lst}{A list of summary statistics. Each element is a Melody object from \code{melody.get.summary}.}

\item{verbose}{whether to generate a plot for microbial feature overlap among studies. Default is FALSE.}
}
\value{
A Melody object that contains the summary statistics over multiple studies participating the meta-analysis.
\item{summary.stat.study}{A list includes summary statistics for each study.}
If verbose=TRUE, generate a plot for microbial feature overlap among studies.
}
\description{
This function takes a list of Melody objects generated by the function melody.get.summary,
align summary statistics by feature IDs, and output a Melody object containing summary statistics of all particpating studies.
The output can be directly used by the function melody.meta.summary.
}
\examples{
\donttest{

data("CRC_abd")
data("meta")

########## Generate summary statistics for study FR ##########
meta_FR <- meta[meta$Study == "FR-CRC",]
CRC_abd_FR <- CRC_abd[,meta_FR$Sample_ID]

sumstats_FR <- melody.get.summary(rel.abd = CRC_abd_FR,
                                  sample.data = meta_FR,
                                  sample.id = "Sample_ID",
                                  study = "Study",
                                  disease = "Group",
                                  verbose = TRUE)

########## Generate summary statistics for study DE ##########
meta_DE <- meta[meta$Study == "DE-CRC",]
CRC_abd_DE <- CRC_abd[,meta_DE$Sample_ID]

sumstats_DE <- melody.get.summary(rel.abd = CRC_abd_DE,
                                  sample.data = meta_DE,
                                  sample.id = "Sample_ID",
                                  study = "Study",
                                  disease = "Group",
                                  verbose = TRUE)

########## Merge summary statistics ##########
sumstats_merge <- melody.merge.summary(list(sumstats_FR, sumstats_DE))
}

}
\seealso{
\code{\link{melody}},
\code{\link{melody.get.summary}},
\code{\link{melody.meta.summary}}
}
