% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melody.meta.summary.R
\name{melody.meta.summary}
\alias{melody.meta.summary}
\title{Meta-analyze summary statistics across studies}
\usage{
melody.meta.summary(
  summary.stats,
  tune.path = c("gsection", "sequence"),
  tune.size.sequence = NULL,
  tune.size.range = NULL,
  tune.type = c("BIC", "HBIC", "KBIC", "EBIC"),
  output.best.one = TRUE,
  tol = 0.001,
  NMAX = 20,
  verbose = FALSE
)
}
\arguments{
\item{summary.stats}{The output of function "melody.get.summary".}

\item{...}{See function melody.}
}
\value{
Same output as the function "melody".
}
\description{
This function directly takes the summary statistics output from the "melody.get.summary"
function and combines the summary statistics across studies for selecting microbial signatures associated with each covariate of interest.
}
\examples{
\donttest{
library("miMeta")
data("CRC_data")
CRC_abd <- CRC_data$CRC_abd
CRC_meta <- CRC_data$CRC_meta

########## Generate summary statistics ##########
rel.abd <- list()
covariate.interest <- list()
for(d in unique(CRC_meta$Study)){
  rel.abd[[d]] <- CRC_abd[CRC_meta$Sample_ID[CRC_meta$Study == d],]
  disease <- as.numeric(CRC_meta$Group[CRC_meta$Study == d] == "CRC")
  names(disease) <- CRC_meta$Sample_ID[CRC_meta$Study == d]
  covariate.interest[[d]] <- data.frame(disease = disease)
}

null.obj <- melody.null.model(rel.abd = rel.abd)

summary.stats <- melody.get.summary(null.obj = null.obj, covariate.interest = covariate.interest)

########## Meta-analysis ##########
meta.result <- melody.meta.summary(summary.stats = summary.stats)
}

}
\seealso{
\code{\link{melody.null.model}},
\code{\link{melody.get.summary}},
\code{\link{melody}}
}
