% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melody.null.model.R
\name{melody.null.model}
\alias{melody.null.model}
\title{Get information from the null model.}
\usage{
melody.null.model(
  rel.abd,
  covariate.adjust = NULL,
  depth.filter = 0,
  prev.filter = 0.1,
  ref = NULL,
  parallel.core = NULL
)
}
\arguments{
\item{...}{See function melody.}
}
\value{
Output a list with each component for a study. The component includes the following elements.
\item{ref}{Reference feature ID for the study.}
\item{p}{A matrix of the estimated mean microbial feature proportions for the study.}
\item{res}{A matrix of the residuals of the relative abundance for the study.}
\item{N}{A vector of microbiome sequencing depth for the study.}
\item{X}{A data frame of the cleaned covariate.adjust for the study.}
\item{rm.sample.idx}{The index of the removed samples for the study.}
}
\description{
The "melody.null.model" function computes the estimated mean microbial feature proportions and residuals  of the relative abundance under the null model of
no associations for each study (so this function does not need the covariate.interest information).
The output of this function will be fed into the "melody.get.summary" function to construct summary statistics
for covariates of interest in each study.
}
\examples{
\donttest{
library("miMeta")
data("CRC_data")
CRC_abd <- CRC_data$CRC_abd
CRC_meta <- CRC_data$CRC_meta

########## Generate summary statistics ##########
rel.abd <- list()
for(d in unique(CRC_meta$Study)){
  rel.abd[[d]] <- CRC_abd[CRC_meta$Sample_ID[CRC_meta$Study == d],]
}

null.obj <- melody.null.model(rel.abd = rel.abd, ref = "Coprococcus catus [ref_mOTU_v2_4874]")
}

}
\seealso{
\code{\link{melody.get.summary}},
\code{\link{melody.meta.summary}},
\code{\link{melody}}
}
