% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{colorectal cancer (CRC) stool metagenomic studies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 572 rows and 18 columns.
}
\usage{
data(CRC_abd, meta)
}
\description{
CRC data includes 100 species from 5 different studies
}
\examples{
data(CRC_abd, meta)
\donttest{iplotCurves(phe, times)}
}
\references{
Wirbel, Jakob et al.
“Meta-analysis of fecal metagenomes reveals global microbial signatures that are specific for colorectal cancer.”
Nature medicine vol. 25,4 (2019): 679-689. doi:10.1038/s41591-019-0406-6
}
\keyword{datasets}
