% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melody.meta.summary.R
\name{melody.meta.summary}
\alias{melody.meta.summary}
\title{Meta-analyze summary statistics across studies}
\usage{
melody.meta.summary(
  Melody,
  tune.path = c("gsection", "sequence"),
  tune.size.sequence = NULL,
  tune.size.range = NULL,
  tune.type = c("BIC", "HBIC", "KBIC", "EBIC"),
  ouput.best.one = TRUE,
  tol = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{Melody}{Melody object.}

\item{...}{Same arguments as the function melody}
}
\value{
Same output as the function melody
}
\description{
This function takes a Melody object that contains summary statistics of all studies participating the
meta-analysis, then combines these summary statistics to select microbial signatures (see the Description of the function melody)
}
\examples{
\donttest{

data("CRC_abd")
data("meta")

########## Generate summary statistics for study FR ##########
meta_FR <- meta[meta$Study == "FR-CRC",]
CRC_abd_FR <- CRC_abd[,meta_FR$Sample_ID]

sumstats_FR <- melody.get.summary(rel.abd = CRC_abd_FR,
                                 sample.data = meta_FR,
                                 sample.id = "Sample_ID",
                                 study = "Study",
                                 disease = "Group",
                                 verbose = TRUE)

########## Generate summary statistics for study DE ##########
meta_DE <- meta[meta$Study == "DE-CRC",]
CRC_abd_DE <- CRC_abd[,meta_DE$Sample_ID]

sumstats_DE <- melody.get.summary(rel.abd = CRC_abd_DE,
                                  sample.data = meta_DE,
                                  sample.id = "Sample_ID",
                                  study = "Study",
                                  disease = "Group",
                                  verbose = TRUE)

########## Merge summary statistics ##########
sumstats_merge <- melody.merge.summary(list(sumstats_FR, sumstats_DE))

########## Meta-analysis ##########
Melody.model <- melody.meta.summary(Melody = sumstats_merge)
}

}
\seealso{
\code{\link{melody}},
\code{\link{melody.get.summary}},
\code{\link{melody.merge.summary}}
}
